splitted tests to conversion (roundtrip) and simple import
